requirements.txt
streamlit
reportlab
python-docx
pandas
lxml
plotly
kaleido
